Commit changes
udpdate the uders and manager users pages and fixed all the conflicts
