+++ pages/dashboard/users/[id].js
@@ -1,411 +1,39 @@
-import React, { useState, useEffect } from 'react';
-import { doc, getDoc, updateDoc, arrayUnion } from 'firebase/firestore';
-import db from '../../../lib/firebase';
+// pages/dashboard/users/[id].js
 import { useRouter } from 'next/router';
-import DashboardLayout from '@/components/Dashboard/DashboardLayout';
-import { AnimatePresence } from 'framer-motion'; // Assuming AnimatePresence is from framer-motion
+import DashboardLayout from '@/components/Dashboard/DashboardLayout'; // Assuming this path is correct
+import UserDetailsContent from '../../../components/Dashboard/UserDetailsContent'; // <--- CORRECTED PATH HERE
 
-// Assuming you have a Notification component like this:
-// import Notification from '@/components/Notification'; // Adjust path as needed
-// For demonstration, I'll assume Notification is a simple functional component.
-// In a real app, you'd import your actual Notification component.
-const Notification = ({ type, message, onClose, show }) => {
-    if (!show) return null;
-
-    const bgColor = type === 'success' ? 'bg-green-500' : 'bg-red-500';
-    return (
-        <div className={`fixed bottom-4 right-4 p-4 rounded-lg shadow-lg text-white ${bgColor} z-[100]`}>
-            <div className="flex justify-between items-center">
-                <span>{message}</span>
-                <button onClick={onClose} className="ml-4 font-bold">X</button>
-            </div>
-        </div>
-    );
-};
-
-
-export default function UserDetails() {
+/**
+ * UserDetailsPage is a Next.js dynamic page for displaying individual user details.
+ * It extracts the user ID from the URL and passes it to the UserDetailsContent component.
+ */
+export default function UserDetailsPage() {
     const router = useRouter();
+    // Destructure 'id' from router.query. This 'id' corresponds to the '[id]' in the file name.
     const { id } = router.query;
 
-    const [name, setName] = useState('');
-    const [email, setEmail] = useState('');
-    // const [password, setPassword] = useState(''); // Not used, consider removing if not needed for admin actions
-    const [phone, setPhone] = useState('');
-    const [photoURL, setPhotoURL] = useState('');
-    const [amount, setAmount] = useState('');
-    const [status, setStatus] = useState('active');
-    const [walletBalance, setWalletBalance] = useState(0);
-    const [earningStatus, setEarningStatus] = useState('running');
-    const [adminNote, setAdminNote] = useState('');
-    const [transactionHistory, setTransactionHistory] = useState([]);
-    const [kycDocs, setKycDocs] = useState([]);
-
-    const [actionType, setActionType] = useState(null); // 'credit' or 'debit'
-
-    // States for loading and notifications
-    const [loading, setLoading] = useState(false);
-    const [showNotification, setShowNotification] = useState(false);
-    const [notificationType, setNotificationType] = useState(''); // 'success' or 'error'
-    const [notificationMessage, setNotificationMessage] = useState('');
-
-    useEffect(() => {
-        if (!id) return;
-
-        const fetchUser = async () => {
-            setLoading(true); // Start loading when fetching user data
-            try {
-                const userRef = doc(db, 'users', id);
-                const userSnap = await getDoc(userRef);
-
-                if (userSnap.exists()) {
-                    const data = userSnap.data();
-                    setName(data.name || '');
-                    setEmail(data.email || '');
-                    setPhone(data.phone || '');
-                    setPhotoURL(data.photoURL || '');
-                    setWalletBalance(data.walletBalance || 0);
-                    setStatus(data.status || 'active');
-                    setEarningStatus(data.earningStatus || 'running');
-                    setAdminNote(data.adminNote || '');
-                    setTransactionHistory(data.transactionHistory || []);
-                    setKycDocs(data.kycDocs || []);
-                } else {
-                    setNotificationType('error');
-                    setNotificationMessage('User not found.');
-                    setShowNotification(true);
-                }
-            } catch (error) {
-                console.error("Error fetching user:", error);
-                setNotificationType('error');
-                setNotificationMessage('Failed to load user details.');
-                setShowNotification(true);
-            } finally {
-                setLoading(false); // End loading regardless of success or failure
-            }
-        };
-
-        fetchUser();
-    }, [id]);
-
-    const handleAction = async (type) => {
-        if (!amount || isNaN(amount) || parseFloat(amount) <= 0) {
-            setNotificationType('error');
-            setNotificationMessage('Please enter a valid amount greater than zero.');
-            setShowNotification(true);
-            return;
-        }
-
-        const value = parseFloat(amount);
-        if (type === 'debit' && value > walletBalance) {
-            setNotificationType('error');
-            setNotificationMessage('Cannot debit more than wallet balance.');
-            setShowNotification(true);
-            return;
-        }
-
-        setLoading(true); // Start loading for action
-        const userRef = doc(db, 'users', id);
-        const date = new Date();
-        const formattedDate = date.toLocaleString('en-US', { // Changed locale for date formatting if needed
-            timeZone: 'America/New_York', // Example time zone for US
-            year: 'numeric',
-            month: 'short',
-            day: 'numeric',
-            hour: '2-digit',
-            minute: '2-digit'
-        });
-
-        const newBalance = type === 'credit' ? walletBalance + value : walletBalance - value;
-
-        try {
-            await updateDoc(userRef, {
-                walletBalance: newBalance,
-                transactionHistory: arrayUnion({
-                    type: type === 'credit' ? 'Credit' : 'Debit',
-                    amount: value,
-                    date: date.toISOString(),
-                    formattedDate
-                })
-            });
-
-            setWalletBalance(newBalance);
-            setTransactionHistory(prev => [...prev, {
-                type: type === 'credit' ? 'Credit' : 'Debit',
-                amount: value,
-                date: date.toISOString(),
-                formattedDate
-            }]);
-            setAmount('');
-            setActionType(null);
-            setNotificationType('success');
-            setNotificationMessage(`${type === 'credit' ? 'Credited' : 'Debited'} successfully!`);
-            setShowNotification(true);
-        } catch (err) {
-            console.error(err);
-            setNotificationType('error');
-            setNotificationMessage(`Failed to ${type} user.`);
-            setShowNotification(true);
-        } finally {
-            setLoading(false); // End loading
-        }
-    };
-
-    const toggleUserStatus = async (newStatus) => {
-        setLoading(true); // Start loading for status change
-        const userRef = doc(db, 'users', id);
-        try {
-            await updateDoc(userRef, { status: newStatus });
-            setStatus(newStatus);
-            setNotificationType('success');
-            setNotificationMessage(`User ${newStatus === 'active' ? 'activated' : 'suspended'} successfully!`);
-            setShowNotification(true);
-        } catch (error) {
-            console.error(error);
-            setNotificationType('error');
-            setNotificationMessage('Failed to update user status.');
-            setShowNotification(true);
-        } finally {
-            setLoading(false); // End loading
-        }
-    };
-
-    const handleAdminNoteSave = async () => {
-        setLoading(true); // Start loading for note save
-        const userRef = doc(db, 'users', id);
-        try {
-            await updateDoc(userRef, { adminNote });
-            setNotificationType('success');
-            setNotificationMessage('Admin note saved successfully!');
-            setShowNotification(true);
-        } catch (error) {
-            console.error(error);
-            setNotificationType('error');
-            setNotificationMessage('Failed to save admin note.');
-            setShowNotification(true);
-        } finally {
-            setLoading(false); // End loading
-        }
-    };
-
-    const handleEarningStatusChange = async (newStatus) => {
-        setLoading(true); // Start loading for earning status change
-        const userRef = doc(db, 'users', id);
-        try {
-            await updateDoc(userRef, { earningStatus: newStatus });
-            setEarningStatus(newStatus);
-            setNotificationType('success');
-            setNotificationMessage(`Earnings ${newStatus === 'running' ? 'resumed' : newStatus === 'paused' ? 'paused' : 'stopped'} successfully!`);
-            setShowNotification(true);
-        } catch (error) {
-            console.error(error);
-            setNotificationType('error');
-            setNotificationMessage('Failed to update earnings status.');
-            setShowNotification(true);
-        } finally {
-            setLoading(false); // End loading
-        }
-    };
-
-    const totalCredit = transactionHistory.filter(txn => txn.type === 'Credit').reduce((acc, txn) => acc + txn.amount, 0);
-    const totalDebit = transactionHistory.filter(txn => txn.type === 'Debit').reduce((acc, txn) => acc + txn.amount, 0);
+    // Show a basic loading state while the ID is being resolved
+    if (!id) {
+        return (
+            <DashboardLayout>
+                <div className="flex items-center justify-center h-screen text-gray-500">
+                    Loading user details...
+                </div>
+            </DashboardLayout>
+        );
+    }
 
     return (
         <DashboardLayout>
-            <AnimatePresence>
-                {loading && (
-                    <div className="absolute inset-0 bg-gray-100 bg-opacity-75 flex items-center justify-center z-50 rounded-2xl">
-                        <div className="w-8 lg:w-12 h-8 lg:h-12 rounded-full animate-spin border-4 border-t-transparent border-l-transparent border-r-blue-500 border-b-purple-500 shadow-lg"></div>
-                    </div>
-                )}
-            </AnimatePresence>
-
-            {showNotification && (
-                <Notification
-                    type={notificationType}
-                    message={notificationMessage}
-                    onClose={() => setShowNotification(false)}
-                    show={showNotification} // Pass show state directly
-                />
-            )}
-
-            <div className="p-4 sm:p-6 lg:p-8 space-y-6">
-                <button onClick={() => router.back()} className="text-sm text-blue-600 hover:underline">← Back to Users</button>
-
-                {/* Quick Wallet Actions */}
-                <div className="bg-white rounded-2xl shadow p-4 mb-4">
-                    <h2 className="text-base font-semibold mb-3">Quick Wallet Actions</h2>
-                    <div className="flex gap-4">
-                        <button
-                            onClick={() => setActionType('credit')}
-                            className="bg-green-600 hover:bg-green-700 text-white text-sm px-4 py-1.5 rounded-lg"
-                            disabled={loading} // Disable buttons during loading
-                        >
-                            Credit
-                        </button>
-                        <button
-                            onClick={() => setActionType('debit')}
-                            className="bg-red-600 hover:bg-red-700 text-white text-sm px-4 py-1.5 rounded-lg"
-                            disabled={loading} // Disable buttons during loading
-                        >
-                            Debit
-                        </button>
-                    </div>
-
-                    {actionType && (
-                        <div className="mt-3 flex flex-col sm:flex-row items-start sm:items-center gap-2">
-                            <input
-                                type="number"
-                                value={amount}
-                                onChange={(e) => setAmount(e.target.value)}
-                                className="border border-gray-300 rounded-lg px-3 py-1.5 text-sm w-full sm:w-auto"
-                                placeholder="Enter amount"
-                                disabled={loading} // Disable input during loading
-                            />
-                            <button
-                                onClick={() => handleAction(actionType)}
-                                className="bg-blue-600 hover:bg-blue-700 text-white text-sm font-medium px-4 py-1.5 rounded-lg"
-                                disabled={loading} // Disable buttons during loading
-                            >
-                                Confirm {actionType === 'credit' ? 'Credit' : 'Debit'}
-                            </button>
-                            <button
-                                onClick={() => { setAmount(''); setActionType(null) }}
-                                className="text-sm text-gray-500 hover:underline"
-                                disabled={loading} // Disable buttons during loading
-                            >
-                                Cancel
-                            </button>
-                        </div>
-                    )}
-                </div>
-
-                {/* User Profile */}
-                <div className="bg-white rounded-2xl shadow p-6 flex flex-col lg:flex-row items-center lg:items-start justify-between gap-6">
-                    <div className="flex flex-col sm:flex-row items-center sm:items-start gap-4">
-                        <img src={photoURL || 'https://via.placeholder.com/100'} className="w-24 h-24 rounded-full border object-cover" alt="User profile" />
-                        <div className="text-center sm:text-left">
-                            <h1 className="text-2xl font-bold">{name}</h1>
-                            <p className="text-sm text-gray-600">ID: {id}</p>
-                            <p className="text-sm text-gray-600">{email}</p>
-                            <p className="text-sm text-gray-600">{phone}</p>
-                            <p className={`text-sm font-semibold ${status === 'active' ? 'text-green-600' : 'text-red-600'}`}>Status: {status}</p>
-                        </div>
-                    </div>
-                    <div className="flex flex-col sm:flex-row gap-2 w-full sm:w-auto">
-                        <button
-                            onClick={() => toggleUserStatus('suspended')}
-                            className="bg-red-600 hover:bg-red-700 text-white text-sm px-4 py-1.5 rounded-lg w-full sm:w-auto"
-                            disabled={loading} // Disable buttons during loading
-                        >
-                            Suspend
-                        </button>
-                        <button
-                            onClick={() => toggleUserStatus('active')}
-                            className="bg-green-600 hover:bg-green-700 text-white text-sm px-4 py-1.5 rounded-lg w-full sm:w-auto"
-                            disabled={loading} // Disable buttons during loading
-                        >
-                            Activate
-                        </button>
-                    </div>
-                </div>
-                <div className="mt-3">
-                    <h3 className="text-sm font-semibold mb-2">Earning Status</h3>
-                    <div className="flex gap-3 flex-wrap">
-                        {[
-                            { label: '✅ Running', value: 'running', color: 'text-green-600' },
-                            { label: '⏸️ Paused', value: 'paused', color: 'text-yellow-600' },
-                            { label: '🛑 Stopped', value: 'stopped', color: 'text-red-600' }
-                        ].map((option) => (
-                            <label key={option.value} className="flex items-center gap-2 cursor-pointer">
-                                <input
-                                    type="radio"
-                                    name="earningStatus"
-                                    value={option.value}
-                                    checked={earningStatus === option.value}
-                                    onChange={() => handleEarningStatusChange(option.value)}
-                                    className="accent-blue-600"
-                                    disabled={loading} // Disable radio buttons during loading
-                                />
-                                <span className={`${option.color} text-sm font-medium`}>{option.label}</span>
-                            </label>
-                        ))}
-                    </div>
-                </div>
-
-                {/* Wallet Balance */}
-                <div className="bg-white rounded-2xl shadow p-6">
-                    <h2 className="text-lg font-semibold mb-1">Wallet Balance</h2>
-                    <p className="text-3xl font-bold">${walletBalance.toLocaleString()}</p>
-                </div>
-
-                {/* Admin Notes */}
-                <div className="bg-white rounded-2xl shadow p-6">
-                    <h2 className="text-lg font-semibold mb-3">Admin Notes</h2>
-                    <textarea
-                        value={adminNote}
-                        onChange={(e) => setAdminNote(e.target.value)}
-                        className="w-full min-h-[120px] border border-gray-300 rounded-xl px-4 py-2 resize-none"
-                        placeholder="Add notes here..."
-                        disabled={loading} // Disable textarea during loading
-                    />
-                    <div className="mt-3 text-right">
-                        <button
-                            onClick={handleAdminNoteSave}
-                            className="bg-blue-600 hover:bg-blue-700 text-white text-sm px-4 py-1.5 rounded-lg"
-                            disabled={loading} // Disable button during loading
-                        >
-                            Save Note
-                        </button>
-                    </div>
-                </div>
-
-                {/* KYC Documents */}
-                <div className="bg-white rounded-2xl shadow p-6">
-                    <h2 className="text-lg font-semibold mb-3">KYC Documents</h2>
-                    {kycDocs.length === 0 ? (
-                        <p className="text-gray-500">No KYC documents uploaded.</p>
-                    ) : (
-                        <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-4">
-                            {kycDocs.map((doc, index) => (
-                                <div key={index} className="border rounded-xl overflow-hidden shadow">
-                                    <img src={doc.url} alt={doc.name} className="w-full h-48 object-cover" />
-                                    <div className="p-3">
-                                        <p className="text-sm font-medium">{doc.name}</p>
-                                        <a href={doc.url} target="_blank" rel="noopener noreferrer" className="text-blue-600 text-xs hover:underline">View</a>
-                                    </div>
-                                </div>
-                            ))}
-                        </div>
-                    )}
-                </div>
-
-                {/* Transaction History */}
-                <div className="bg-white rounded-2xl shadow p-6">
-                    <h2 className="text-lg font-semibold mb-3">Transaction History</h2>
-                    <div className="flex justify-between text-sm font-medium mb-2">
-                        <p>Total Credit: <span className="text-green-600">${totalCredit.toLocaleString()}</span></p>
-                        <p>Total Debit: <span className="text-red-600">${totalDebit.toLocaleString()}</span></p>
-                    </div>
-                    <ul className="divide-y text-sm">
-                        {transactionHistory.length === 0 ? (
-                            <li className="text-gray-500">No records yet.</li>
-                        ) : (
-                            transactionHistory.map((txn, i) => (
-                                <li key={i} className="flex justify-between py-2 items-center">
-                                    <div>
-                                        <p className="font-medium">{txn.type}</p>
-                                        <p className="text-xs text-gray-500">{txn.formattedDate || txn.date}</p>
-                                    </div>
-                                    <span className={txn.type === 'Debit' ? 'text-red-600' : 'text-green-600'}>
-                                        ${txn.amount.toLocaleString()}
-                                    </span>
-                                </li>
-                            ))
-                        )}
-                    </ul>
-                </div>
-            </div>
+            {/* The "Back to Users" button belongs here as useRouter is available in the Page component */}
+            <button
+                onClick={() => router.back()}
+                className="text-sm text-blue-600 hover:underline p-4" // Added some padding for better click area
+            >
+                ← Back to Users
+            </button>
+            {/* Render the UserDetailsContent component, passing the extracted ID as a prop */}
+            <UserDetailsContent userId={id} />
         </DashboardLayout>
     );
 }
